Adjusting obsolets for the transport definition change
