+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -117,6 +117,7 @@
     <Compile Include="Transports\Msmq\MsmqScopeOptions.cs" />
     <Compile Include="Transports\Msmq\MsmqTransportInfrastructure.cs" />
     <Compile Include="Transports\MulticastTransportOperation.cs" />
+    <Compile Include="Transports\TransportInfrastructure.cs" />
     <Compile Include="Transports\TransportOperations.cs" />
     <Compile Include="Transports\TransportSubscriptionInfrastructure.cs" />
     <Compile Include="Transports\UnicastTransportOperation.cs" />
